requirements.txt
# Malware Detector by Code Analysis - Python Dependencies

# OpenRouter/OpenAI API client
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Standard library dependencies (no installation needed):
# - subprocess
# - pathlib
# - json
# - re
# - shutil
# - datetime

